Model the table and hands
Creates a class to handle the notion of a 'Hand',
which keeps score and allows the player to request
cards from the table.

The table contains the deck, is responsible for shuffling and
dealing cards, and ultimately calculating the result based on
the messages it receives from the hands.

Not yet implemented the dealer's hand, which will behave much
like a regular hand except it must hit until 16, at which point
it should stand